Relocated decoy generation from AbstractScoringTask to DecoyGenerator
* Moved GenerateAllWithCollisionDetection (was GenerateDecoys) and BuildDecoyFromSequence into OspreySharp.Scoring.DecoyGenerator, arithmetic verbatim
* Replaced the ambient PipelineContext logger with an injected Action<string> so Scoring keeps no Tasks/PipelineContext dependency
* Updated the single caller (PerFileScoringTask) and a stale ScoringTest comment
* Stage A of the AbstractScoringTask decomposition; AbstractScoringTask -122 lines, byte-parity + perf gates green

See ai/todos/active/TODO-20260611_ospreysharp_decouple_abstractscoring.md
