+++ pwiz_tools/OspreySharp/OspreySharp/Tasks/AbstractScoringTask.cs
@@ -112,128 +112,6 @@ namespace pwiz.OspreySharp.Tasks
 
 
         /// <summary>
-        /// Generate decoy entries from the target library with collision detection.
-        /// Matches Rust DecoyGenerator.generate_all_with_collision_detection:
-        ///   1. Build set of target sequences (stripped) for collision detection
-        ///   2. For each target, try reversing
-        ///   3. If reversed collides or is palindromic, try cycling with lengths 1..10
-        ///   4. If all methods fail, exclude the target-decoy pair
-        /// Modifies <paramref name="validTargets"/> to contain only targets that
-        /// produced valid decoys (Rust: library = valid_targets; library.extend(decoys)).
-        /// </summary>
-        protected List<LibraryEntry> GenerateDecoys(
-            List<LibraryEntry> targets, OspreyConfig config,
-            out List<LibraryEntry> validTargets)
-        {
-            _ctx.LogInfo(string.Format("Generating decoys using {0} method...", config.DecoyMethod));
-
-            // Build set of all target (stripped) sequences for collision detection.
-            var targetSequences = new HashSet<string>(StringComparer.Ordinal);
-            foreach (var t in targets)
-            {
-                if (!t.IsDecoy)
-                    targetSequences.Add(t.Sequence);
-            }
-
-            // Generate decoys in parallel (matches Rust's par_iter approach).
-            // Each target produces a (target, decoy) pair or is excluded.
-            int nReversed = 0, nCycled = 0, nExcluded = 0, nSkipped = 0;
-            var results = new (LibraryEntry target, LibraryEntry decoy, int kind)[targets.Count];
-            // kind: 0=skip, 1=reversed, 2=cycled, 3=excluded
-
-            Parallel.For(0, targets.Count, i =>
-            {
-                var target = targets[i];
-                if (target.IsDecoy || target.Fragments == null || target.Fragments.Count == 0)
-                {
-                    results[i] = (null, null, 0);
-                    return;
-                }
-
-                // Each thread gets its own generator (DecoyGenerator is lightweight)
-                var gen = new DecoyGenerator();
-                int[] mapping;
-                string reversedSeq = gen.ReverseSequence(target.Sequence, out mapping);
-
-                if (reversedSeq != target.Sequence && !targetSequences.Contains(reversedSeq))
-                {
-                    var decoy = BuildDecoyFromSequence(target, reversedSeq, mapping);
-                    if (decoy != null)
-                    {
-                        results[i] = (target, decoy, 1);
-                        return;
-                    }
-                }
-
-                // Fallback: cycling with lengths 1..min(len, 10)
-                int maxRetries = Math.Min(target.Sequence.Length, 10);
-                for (int cycleLength = 1; cycleLength <= maxRetries; cycleLength++)
-                {
-                    string cycledSeq = gen.CycleSequence(target.Sequence, cycleLength, out mapping);
-                    if (cycledSeq != target.Sequence && !targetSequences.Contains(cycledSeq))
-                    {
-                        var decoy = BuildDecoyFromSequence(target, cycledSeq, mapping);
-                        if (decoy != null)
-                        {
-                            results[i] = (target, decoy, 2);
-                            return;
-                        }
-                    }
-                }
-
-                results[i] = (null, null, 3);
-            });
-
-            // Collect results (sequential, preserves order)
-            validTargets = new List<LibraryEntry>(targets.Count);
-            var decoys = new List<LibraryEntry>(targets.Count);
-            foreach (var r in results)
-            {
-                switch (r.kind)
-                {
-                    case 0: nSkipped++; break;
-                    case 1: nReversed++; validTargets.Add(r.target); decoys.Add(r.decoy); break;
-                    case 2: nCycled++; validTargets.Add(r.target); decoys.Add(r.decoy); break;
-                    case 3: nExcluded++; break;
-                }
-            }
-
-            _ctx.LogInfo(string.Format(
-                "Generated {0} decoys from {1} targets ({2} excluded due to collisions)",
-                decoys.Count, targets.Count, nExcluded));
-            return decoys;
-        }
-
-
-        /// <summary>
-        /// Build a decoy LibraryEntry from a decoy sequence and position mapping.
-        /// Mirrors DecoyGenerator.Generate's construction but takes an already-chosen sequence.
-        /// </summary>
-        private static LibraryEntry BuildDecoyFromSequence(
-            LibraryEntry target, string decoySequence, int[] positionMapping)
-        {
-            var decoy = new LibraryEntry(
-                target.Id | 0x80000000u,
-                decoySequence,
-                "DECOY_" + target.ModifiedSequence,
-                target.Charge,
-                target.PrecursorMz,
-                target.RetentionTime);
-            decoy.RtCalibrated = target.RtCalibrated;
-            decoy.IsDecoy = true;
-            decoy.Modifications = DecoyGenerator.RemapModificationsStatic(
-                target.Modifications, positionMapping);
-            decoy.Fragments = DecoyGenerator.RecalculateFragmentsStatic(
-                target, positionMapping, decoySequence);
-            decoy.ProteinIds = new List<string>();
-            foreach (string p in target.ProteinIds)
-                decoy.ProteinIds.Add("DECOY_" + p);
-            decoy.GeneNames = new List<string>(target.GeneNames);
-            return decoy;
-        }
-
-
-        /// <summary>
         /// Extract unique isolation windows from the first cycle of MS2 spectra.
         /// </summary>
         protected List<IsolationWindow> ExtractIsolationWindows(List<Spectrum> spectra)
